Win uses %I64d, Nix uses %lld.

+++ makeself.c
@@ -152,7 +152,7 @@ static void build_info_hdr(void)
 			app_type = 6;
 			break;
 		default:
-			fail("something that should never fail failed.");	
+			fail("something that should never fail failed.");
 	}
 
 	wbe64(info_header + 0x00, auth_id);
@@ -192,7 +192,7 @@ static void build_sec_hdr(void)
 		wbe32(ptr + 0x14, 0);		// unknown
 		wbe32(ptr + 0x18, 0);		// unknown
 
-		if (phdr[i].p_type == 1) 
+		if (phdr[i].p_type == 1)
 			wbe32(ptr + 0x1c, 1);	// encrypted LOAD phdr
 		else
 			wbe32(ptr + 0x1c, 0);	// no loadable phdr
@@ -259,7 +259,7 @@ static void build_meta_hdr(void)
 
 	// add encrypted phdr information
 	for (i = 0; i < ehdr.e_phnum; i++) {
-		if (phdr[i].p_type == 1) 
+		if (phdr[i].p_type == 1)
 			meta_add_load(ptr, i);
 		else
 			meta_add_phdr(ptr, i);
@@ -280,14 +280,14 @@ static void calculate_hashes(void)
 
 	for (i = 0; i < ehdr.e_phnum; i++) {
 		printf("%d\n", i);
-		printf("%08x\n", phdr[i].p_off);
+		printf("%08x\n", (unsigned int)phdr[i].p_off);
 		memset(keys + (i * 8 * 0x10), 0, 0x20);
 		sha1_hmac(keys + ((i * 8) + 2) * 0x10,
 		          elf + phdr[i].p_off,
 			  phdr[i].p_filesz,
 			  keys + (i * 8) * 0x10
 			 );
-	}	
+	}
 }
 
 static void build_hdr(void)
@@ -364,7 +364,7 @@ static void get_version(const char *v)
 			ptr++;
 			continue;
 		}
-	
+
 		fprintf(stderr, "WARNING: invalid sdk_version, using 1.0.0\n");
 		version = 1ULL << 48;
 		return;
@@ -418,8 +418,8 @@ int main(int argc, char *argv[])
 	ctrl_offset = round_up(version_offset + 0x10, ALIGNMENT);
 	sec_offset = round_up(ctrl_offset + 0x70, ALIGNMENT);
 	elf_offset = round_up(sec_offset + ehdr.e_phnum * 0x20, ALIGNMENT);
-	phdr_offset = round_up(elf_offset + ehdr.e_ehsize, ALIGNMENT);	
-	shdr_offset = round_up(phdr_offset + ehdr.e_phentsize * ehdr.e_phnum, ALIGNMENT);	
+	phdr_offset = round_up(elf_offset + ehdr.e_ehsize, ALIGNMENT);
+	shdr_offset = round_up(phdr_offset + ehdr.e_phentsize * ehdr.e_phnum, ALIGNMENT);
 	meta_offset = round_up(shdr_offset + ehdr.e_shentsize * ehdr.e_shnum, ALIGNMENT);
 	header_size = round_up(meta_offset + 0x80 + ehdr.e_phnum * (0x30 + 0x20 + 0x60) + 0x30, 0x80);
